+++ src/Models/ReportJob.php
@@ -3,6 +3,7 @@
 namespace MBLSolutions\Report\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
 class ReportJob extends Model
@@ -17,5 +18,16 @@ class ReportJob extends Model
 
     /** {@inheritDoc} */
     protected $guarded = [];
+
+
+    protected $casts = [
+        'processed' => 'integer',
+        'total' => 'integer',
+    ];
+
+    public function report(): BelongsTo
+    {
+        return $this->belongsTo(Report::class);
+    }
 
 }